backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.40
alembic==1.15.2
python-dotenv==1.1.0
requests==2.32.3
geopandas==1.0.1
shapely==2.1.0
pyproj==3.7.1
openpyxl==3.1.5
pandas==2.2.3
numpy==2.2.4
